Move the permissions call out of api_controller
