+++ app/views/api/permissions/show.builder
